+++ doc/whatsnew.rst
@@ -1,8 +1,10 @@
 What's new
 **********
 
-.. Next release
-.. ============
+Next release
+============
+
+- Bugfix: genno 1.23.0 would return :class:`.AttrSeries` with :class:`pandas.Index` (instead of 1-D :class:`pandas.MultiIndex`) from :meth:`~.AttrSeries.squeeze` (:issue:`120`, :pull:`122`).
 
 v1.23.0 (2024-01-26)
 ====================